Picking up and putting down treasure
Added "treasure" objects, which can be picked up if you switch to inventory slot 2 (press 2) and left click on it. Left click again to place it down in front of you.

The treasure object rotates to match the player when placed down, and also has gravity that only works when it "exists" within the world.