Added haste potion and shop
Added shop with potions on it. Buying a potion will put the potion in the earliest slot that can hold it. Using the potion (with right click or E) will consume the potion, returning it to the location on the shop, and give you a "haste" effect for 12 seconds, which turns the stamina bar yellow and instead shows the time remaining for the effect.